requirements.txt
# Core dependencies
websockets==12.0
python-dotenv==1.0.0
loguru==0.7.2
openai==1.54.0
httpx<0.28.0
requests==2.31.0

# Optional for testing
pytest==7.4.3
pytest-asyncio==0.21.1
